requirements.txt
asgiref==3.5.2
attrs==22.1.0
cached-property==1.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0.6
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
environs==9.5.0
gunicorn==20.1.0
idna==3.3
ippanel==1.1.0
isodate==0.6.1
lxml==4.9.1
marshmallow==3.17.0
packaging==21.3
Pillow==9.2.0
platformdirs==2.5.2
psycopg==3.0.16
PyJWT==2.4.0
pyparsing==3.0.9
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.1
requests-file==1.5.1
requests-toolbelt==0.9.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
Unidecode==1.3.4
urllib3==1.26.11
whitenoise==6.2.0
zeep==4.1.0
